+++ Breaking/app/src/main/java/com/dope/breaking/model/request/RequestSignUp.kt
@@ -1,4 +1,4 @@
-package com.dope.breaking.model
+package com.dope.breaking.model.request
 
 import org.json.JSONObject
 
@@ -18,10 +18,9 @@ data class RequestSignUp(
 
     /**
      * 모든 필드 값들을 json 데이터로 변환해주는 메소드
-     * @param - None
-     * @return - JSONObject: 모든 필드들을 json 으로 변환한 객체 반환
-     * @author - Seunggun Sin
-     * @since - 2022-07-08
+     * @return JSONObject: 모든 필드들을 json 으로 변환한 객체 반환
+     * @author Seunggun Sin
+     * @since 2022-07-08
      */
     fun convertFieldsToJson(): JSONObject {
         val jsonObject = JSONObject()
@@ -37,10 +36,9 @@ data class RequestSignUp(
 
     /**
      * json 형태의 데이터를 json string 형태로 변환해주는 메소드
-     * @param - None
-     * @return - String: 필드 값들을 json 형태로 변환 후 뒤, 문자열로 변환한 값 반환
-     * @author - Seunggun Sin
-     * @since - 2022-07-08
+     * @return String: 필드 값들을 json 형태로 변환 후 뒤, 문자열로 변환한 값 반환
+     * @author Seunggun Sin
+     * @since 2022-07-08
      */
     fun convertJsonToString(): String {
         val json = convertFieldsToJson() // json 으로 변환
